requirements.txt
streamlit==0.72.0
numpy==1.17.4
requests==2.24.0
altair==4.1.0
pandas==1.1.5
PyYAML==5.4.1
